Added input controller.
Resolved bug in FPS and IPS limiting logic that caused actual performance to be limited slightly more than intended.  A cap of 120 should now produce much closer to 120, instead of the buggy over or under limiting that occurred before.

+++ include/se/graphics/graphicsController.hpp
@@ -71,7 +71,7 @@ namespace se::graphics {
              * 
              *  Negative values disable FPS cap.
              */
-            int target_frame_time = -1;
+            int64_t target_frame_time = -1;
 
             /*!
              *  Graphics Tasks.
